Added an Nullary constructor optimisation to diabetes.
Now we will translate nullary constructors (Nil, True, False, Zero)
to heap variables ($Nil, $True,.. )

So if you have something like:
  let {x = CON ( FOO )} in x;

it will now be translated to:
  $FOO

where $FOO is defined as:
$FOO = CON( FOO );

at the toplevel. An implication of this change is that Constructors
must always be saturated, i.e. you can't partial apply them like you
can in Haskell. And since we don't know the arity of constructors we
can't eta-expand to solve this.