Chars and Strings added, SugarTree, AST, Pretty...
Chars 'x' and String "Hi there!" is now added and seems to work
prefectly. Strings are just a list of chars and chars are boxed (C#).

We need to add primtive ops to chars, I recommend chr and ord but maybe
different names. Then we can piggy-back on ints for the most part.